+++ test/unit/event_machine_proxy_test.rb
@@ -24,6 +24,14 @@ describe ExceptionalSynchrony::EventMachineProxy do
     end
   end
 
+  def stop_em_after_defers_finish!(em)
+    check_finished_counter = 0
+    em.add_periodic_timer(0.1) do
+      (check_finished_counter += 1) > 20 and raise "defer never finished!"
+      em.defers_finished? and em.stop
+    end
+  end
+
   before do
     @em = ExceptionalSynchrony::EventMachineProxy.new(EventMachine, nil)
     @yielded_value = nil
@@ -56,6 +64,14 @@ describe ExceptionalSynchrony::EventMachineProxy do
     @em.stop
   end
 
+  it "should set thread variable :em_synchrony_reactor_thread running to false when stop" do
+    @em.run do
+      assert_equal true, Thread.current.thread_variable_get(:em_synchrony_reactor_thread)
+      @em.stop
+      assert_equal false, Thread.current.thread_variable_get(:em_synchrony_reactor_thread)
+    end
+  end
+
   it "should proxy connect" do
     ServerClass = Class.new
     mock(EventMachine).connect(ServerClass, 8080, :handler, :extra_arg).yields(:called)
@@ -78,18 +94,41 @@ describe ExceptionalSynchrony::EventMachineProxy do
   end
 
   describe "#defer" do
-    it "should output its block's output when it doesn't raise an error" do
-      ExceptionHandling.logger = Logger.new(STDERR)
+    before do
+      logger = Logger.new(STDERR)
+      logger.extend ContextualLogger::LoggerMixin
+      ExceptionHandling.logger = logger
+    end
 
+    it "should output its block's output when it doesn't raise an error, by default" do
       @em.run do
         assert_equal 12, @em.defer("#defer success") { 12 }
         @em.stop
       end
     end
 
-    it "should raise an error when its block raises an error" do
-      ExceptionHandling.logger = Logger.new(STDERR)
+    it "should not wait for its block to run if option is passed" do
+      @block_ran = false
 
+      @em.run do
+        assert_nil @em.defer("#defer success", wait_for_result: false) { @block_ran = true; 12 }
+        refute @block_ran
+        stop_em_after_defers_finish!(@em)
+      end
+
+      assert @block_ran
+    end
+
+    it "should handle exceptions when not waiting for its block to run" do
+      mock(ExceptionHandling).log_error(is_a(RuntimeError), "defer", {})
+
+      @em.run do
+        assert_nil @em.defer("#defer success", wait_for_result: false) { raise RuntimeError, "error in defer" }
+        stop_em_after_defers_finish!(@em)
+      end
+    end
+
+    it "should raise an error when its block raises an error" do
       @em.run do
         ex = assert_raises(ArgumentError) do
           @em.defer("#defer raising an error") { raise ArgumentError, "!!!" }
@@ -140,12 +179,25 @@ describe ExceptionalSynchrony::EventMachineProxy do
       end
 
       describe "without error" do
+        before do
+          Thread.current.thread_variable_set(:em_synchrony_reactor_thread, nil)
+        end
+
         [:log, :raise].each do |on_error|
           describe "when using #{method} and on_error = #{on_error}" do
             it "should dispatch to the proxy's synchrony method instead of run iff synchrony" do
               dispatched = false
               assert_equal method, (@proxy.run(on_error: on_error) { dispatched = true })
               assert_equal true, dispatched
+            end
+
+            if method == :synchrony
+              it "should set thread variable :em_synchrony_reactor_thread running to true" do
+                assert_nil Thread.current.thread_variable_get(:em_synchrony_reactor_thread)
+                @proxy.run(on_error: on_error) do
+                  assert_equal true, Thread.current.thread_variable_get(:em_synchrony_reactor_thread)
+                end
+              end
             end
           end
         end